refactor: rename file to follow project pattern
